tests/wpt/web-platform-tests/html/semantics/forms/the-input-element/input-stepup-weekmonth.html

<!DOCTYPE HTML>
<html>
<title>Forms</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<h3>input_stepUp</h3>
<input type="month" id="month_input" max="2009-02" step="1" value="2010-02">
<input type="week" id="week_input" max="2009-W02" step="1" value="2010-W02">

<script>
    function testStepUpOverflow(id, value, type) {
      test(function() {
        var input = document.getElementById(id);
        input.stepUp();
        assert_equals(input.value, value, "value shouldn't change.");
      }, "Calling stepUp() on input -" + type + "- where value > max should not modify value.");
    }

    testStepUpOverflow("month_input", "2010-02", "month");
    testStepUpOverflow("week_input", "2010-W02", "week");
</script>
</html>
